changed variable type for handling money to numeric 16,2

+++ BankOfSPARSA/scripts/bankofsparsa.sql
@@ -17,7 +17,7 @@ ALTER TABLE login ADD UNIQUE (emailAddr);
 CREATE TABLE accounts (userID int AUTO_INCREMENT,
   accountNum bigint(10) unsigned,
   accountPIN int(4),
-  balance float(20,2) default 0.0,
+  balance numeric(16,2) default 0.0,
   primary key (userID));
 ALTER TABLE accounts ADD UNIQUE (accountNum);
 
@@ -27,7 +27,7 @@ CREATE TABLE transactions ( transactionID INT NOT NULL AUTO_INCREMENT,
   src_acct BIGINT UNSIGNED NOT NULL,
   dst_routing_num BIGINT UNSIGNED NOT NULL,
   dst_acct BIGINT UNSIGNED NOT NULL,
-  amount FLOAT NULL DEFAULT NULL,
+  amount numeric(16,2) NULL DEFAULT NULL,
   timestamp DATETIME NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (transactionID));
